incident_management_project/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.10.0
psycopg2==2.9.9
pika==1.1.0
django-filter==23.5.0